% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weightedKM.R
\name{calculate_risk_set}
\alias{calculate_risk_set}
\title{Calculate risk set at specified time points}
\usage{
calculate_risk_set(times, weights, risk_points)
}
\arguments{
\item{times}{Event/censoring times.}

\item{weights}{Weights for each subject.}

\item{risk_points}{Time points for risk set calculation.}
}
\value{
Vector of risk set sizes at each time point.
}
\description{
Calculate risk set at specified time points
}
