% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weightedKM.R
\name{compute_km_group}
\alias{compute_km_group}
\title{Compute KM estimates for a group}
\usage{
compute_km_group(time, event, weights, at.points)
}
\arguments{
\item{time}{Event/censoring times.}

\item{event}{Event indicators.}

\item{weights}{Weights for each subject.}

\item{at.points}{Time points for estimation.}
}
\value{
List with KM survival, standard error, risk set, and event increments.
}
\description{
Compute KM estimates for a group
}
