% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weightedKM.R
\name{format_pval}
\alias{format_pval}
\title{Format p-value for display}
\usage{
format_pval(pval, eps = 0.001, digits = 3)
}
\arguments{
\item{pval}{P-value to format.}

\item{eps}{Threshold for displaying as "<eps".}

\item{digits}{Number of decimal places.}
}
\value{
Formatted p-value string.
}
\description{
Format p-value for display
}
