% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weightedKM.R
\name{get_censor_times}
\alias{get_censor_times}
\title{Get censoring times for a group}
\usage{
get_censor_times(time, event, at.points)
}
\arguments{
\item{time}{Event/censoring times.}

\item{event}{Event indicators.}

\item{at.points}{Time points for estimation.}
}
\value{
List with censoring times and their indices in at.points.
}
\description{
Get censoring times for a group
}
