% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weightedKM.R
\name{Nkm_Weighted}
\alias{Nkm_Weighted}
\title{Weighted number of events up to time x}
\usage{
Nkm_Weighted(x, error, W = rep(1, length(error)))
}
\arguments{
\item{x}{Time point.}

\item{error}{Vector of event/censoring times.}

\item{W}{Weights for each observation (default: all 1).}
}
\value{
Weighted count of events up to time x.
}
\description{
Weighted number of events up to time x
}
