% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weightedKM.R
\name{prepare_group_data}
\alias{prepare_group_data}
\title{Prepare group data (extract vectors)}
\usage{
prepare_group_data(
  df,
  tte.name,
  event.name,
  treat.name,
  weight.name = NULL,
  treat_value = 1
)
}
\arguments{
\item{df}{Data frame.}

\item{tte.name}{Time-to-event column name.}

\item{event.name}{Event indicator column name.}

\item{treat.name}{Treatment group column name.}

\item{weight.name}{Weight column name (optional).}

\item{treat_value}{Value indicating the group (default: 1).}
}
\value{
List with vectors for time, event, and weights.
}
\description{
Prepare group data (extract vectors)
}
