% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weightedKM.R
\name{Rkm_Weighted}
\alias{Rkm_Weighted}
\title{Weighted number at risk at time x}
\usage{
Rkm_Weighted(x, error, W = rep(1, length(error)))
}
\arguments{
\item{x}{Time point.}

\item{error}{Vector of event/censoring times.}

\item{W}{Weights for each observation (default: all 1).}
}
\value{
Weighted count of subjects at risk at time x.
}
\description{
Weighted number at risk at time x
}
