% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weightedKM.R
\name{validate_input}
\alias{validate_input}
\title{Validate input data frame and columns}
\usage{
validate_input(df, required_cols)
}
\arguments{
\item{df}{Data frame.}

\item{required_cols}{Vector of required column names.}
}
\value{
Stops with error if columns are missing; otherwise, returns invisibly.
}
\description{
Validate input data frame and columns
}
